pyproject.toml
[tool.poetry]
name = "buildingmotif"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
rdflib = "^6.1.1"
SQLAlchemy = "^1.4"
pyaml = "^21.10.1"
networkx = "^2.7.1"
types-PyYAML = "^6.0.4"
rdflib-sqlalchemy = {git = "https://github.com/haneslinger/rdflib-sqlalchemy.git", rev = "Take-session-manager"}

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
pylama = "^8.3.7"
pre-commit = "^2.17.0"
pytest-cov = "^3.0.0"
Sphinx = "^4.4.0"
sphinx-rtd-theme = "^1.0.0"
mypy = "^0.931"
sqlalchemy2-stubs = "^0.0.2-alpha.20"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pylint.messages_control]
disable = "C0330, C0326"

[tool.pylint.format]
max-line-length = "88"
